cannot solve developer mode; redis error
