+++ __tests__/actions/randomNumber.js
@@ -5,23 +5,25 @@ const ActionHero = require(path.join(__dirname, '/../../index.js'))
 const actionhero = new ActionHero.Process()
 let api
 
-describe('Action: randomNumber', () => {
-  beforeAll(async () => { api = await actionhero.start() })
-  afterAll(async () => { await actionhero.stop() })
+describe('Action', () => {
+  describe('randomNumber', () => {
+    beforeAll(async () => { api = await actionhero.start() })
+    afterAll(async () => { await actionhero.stop() })
 
-  let firstNumber = null
+    let firstNumber = null
 
-  test('generates random numbers', async () => {
-    let {randomNumber} = await api.specHelper.runAction('randomNumber')
-    expect(randomNumber).toBeGreaterThan(0)
-    expect(randomNumber).toBeLessThan(1)
-    firstNumber = randomNumber
-  })
+    test('generates random numbers', async () => {
+      let {randomNumber} = await api.specHelper.runAction('randomNumber')
+      expect(randomNumber).toBeGreaterThan(0)
+      expect(randomNumber).toBeLessThan(1)
+      firstNumber = randomNumber
+    })
 
-  test('is unique / random', async () => {
-    let {randomNumber} = await api.specHelper.runAction('randomNumber')
-    expect(randomNumber).toBeGreaterThan(0)
-    expect(randomNumber).toBeLessThan(1)
-    expect(randomNumber).not.toEqual(firstNumber)
+    test('is unique / random', async () => {
+      let {randomNumber} = await api.specHelper.runAction('randomNumber')
+      expect(randomNumber).toBeGreaterThan(0)
+      expect(randomNumber).toBeLessThan(1)
+      expect(randomNumber).not.toEqual(firstNumber)
+    })
   })
 })
