+++ __tests__/actions/validationTest.js
@@ -5,17 +5,19 @@ const ActionHero = require(path.join(__dirname, '/../../index.js'))
 const actionhero = new ActionHero.Process()
 let api
 
-describe('Action: Validator', () => {
-  beforeAll(async () => { api = await actionhero.start() })
-  afterAll(async () => { await actionhero.stop() })
+describe('Action', () => {
+  describe('validationTest', () => {
+    beforeAll(async () => { api = await actionhero.start() })
+    afterAll(async () => { await actionhero.stop() })
 
-  test('fails with no params', async () => {
-    let {error} = await api.specHelper.runAction('validationTest', {})
-    expect(error).toEqual('Error: string is a required parameter for this action')
-  })
+    test('fails with no params', async () => {
+      let {error} = await api.specHelper.runAction('validationTest', {})
+      expect(error).toEqual('Error: string is a required parameter for this action')
+    })
 
-  test('fails with a number', async () => {
-    let {error} = await api.specHelper.runAction('validationTest', {string: 87})
-    expect(error).toEqual('Error: Input for parameter "string" failed validation!')
+    test('fails with a number', async () => {
+      let {error} = await api.specHelper.runAction('validationTest', {string: 87})
+      expect(error).toEqual('Error: Input for parameter "string" failed validation!')
+    })
   })
 })
